Bug fix: possible buffer overrun.
After the best chunk is selected, if alignment is request, its size
is rounded down before the allocation is made.  If the requested
allocation size is not a multiple of the alignment size, this
may result is allocation of less space than was requested (buffer
overrun).

This fixes the problem, I think, by rounding the requested allocation
size up to a multiple of the alignment size before selecting
the best chunk.