Ability for DSO extension to support new datatype
Current implementation of rbldnsd specifies that
dataset type must be known at time of compile as
they must be registered by macros.

This patch adds a new native data type: 'extension'
that allows one to create a dynamic shared object
that is able to define the specific dataset type
handlers for:

    - reset
    - start
    - stop
    - query
    - dump

In addition to the standard functions, this patch
also adds the ability to override the loaddataset
function if, for example, the dataset source is a
binary format as opposed to the generic ascii text
line by line read that the native wrapper loading
mechanism provides.

To create an extension simply create a DSO that
defines and exports functions:

    rbldnsd_extension_init

and optionally provide the dataset specific functions:

    rbldnsd_extension_reset
    rbldnsd_extension_start
    rbldnsd_extension_stop
    rbldnsd_extension_query
    rbldnsd_extension_dump

and if you need to override dataset loading:

    rbldnsd_extension_loaddataset

To provide this new functionality, rbldnsd must
be configured with flag:

    --enable-dso

Pre-existing 'extension' support where one only
defines rbldnsd_extension_init, and optionally define
the extension 'hook' functions continues to be supported.

LIMITATIONS

The limitations of this design is that one may only provide
one custom extension dataset type at run time.  One cannot
specify multiple modules to provide multiple different dataset
types on one running daemon.