configure: use a valid program for the connect() library test.
To test if connect() is available (and in what library, if any, it
lives), the ./configure script was compiling and linking

  int main(int argc, char **argv) {
    gethostbyname();
    connect();
    return 0;
  }

This program, however, is invalid and fails to build with pedantic
compilers. The functions gethostbyname() and connect() are undefined;
and once we include the necessary headers to define them, they're
called incorrectly, with the wrong number/type of arguments.

The program is never actually *run*, so we fix this by including the
headers that define gethostbyname() and connect(), and then by
changing those two calls to pass correctly-typed junk arguments to the
functions.

This can be detected by the "implicit-function-declaration" warning in
gcc or clang.